# app_backend/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Inventario, Préstamos y Penalizaciones
  description: >
    Esta API maneja la autenticación de usuarios, administración de perfiles, 
    control centralizado de stock (materiales, herramientas, placas, libros), 
    flujo de préstamos/reservas mediante validación por QR, gestión de penalizaciones 
    y generación de reportes/estadísticas.
  version: 2.0.0
servers:
  - url: /
    description: Servidor Base Local
paths:
...
  /api/usuarios:
    get:
      tags:
        - Usuarios
      summary: Lista los usuarios registrados
      description: Devuelve una lista completa de usuarios. (Solo Admin).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuarios obtenida con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetail'
        '403':
          description: Prohibido. Permisos insuficientes.

    post:
      tags:
        - Usuarios
      summary: Alta (ABM) de un nuevo usuario
      description: Permite registrar un nuevo usuario en el sistema. (Solo Admin).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuario creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: Datos inválidos o conflicto.
        '403':
          description: Prohibido. Permisos insuficientes.
...
  /api/usuarios/{usuario_id}:
    parameters:
      - name: usuario_id
        in: path
        required: true
        description: Identificador único del usuario.
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Usuarios
      summary: Obtiene los detalles de un usuario específico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detalles del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: Usuario no encontrado.

    put:
      tags:
        - Usuarios
      summary: Modifica los datos de un usuario existente
      description: Reemplazo completo de la información general del usuario.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario modificado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: Datos de actualización inválidos.
        '404':
          description: Usuario no encontrado.

    delete:
      tags:
        - Usuarios
      summary: Baja o desactivación lógica de un usuario (Opción 1)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usuario dado de baja con éxito.
        '404':
          description: Usuario no encontrado.
...
  /api/usuarios/{usuario_id}/reservas:
    get:
      tags:
        - Usuarios
      summary: Obtiene el historial personal de préstamos/pedidos
      description: Admin y bibliotecario pueden consultar cualquier usuario; alumno y profesor solo pueden consultar su propio historial.
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Historial obtenido correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanCompact'
        '404':
          description: Alumno no encontrado.
        '403':
          description: Prohibido. El usuario autenticado no puede consultar reservas ajenas.


  # ==========================================
  # 3. Inventario y Materiales (/api/items)
  # ==========================================
  /api/articulos:
    get:
      tags:
        - Inventario y Materiales
      summary: Lista el material disponible
      description: Acceso público (sin autenticación requerida). Incluye soporte para filtros por tipo, estado y disponibilidad.
      parameters:
        - name: category
          in: query
          required: false
          description: Filtrar por tipo o categoría.
          schema:
            type: string
        - name: condition
          in: query
          required: false
          description: Filtrar por condición física.
          schema:
            type: string
        - name: available
          in: query
          required: false
          description: Filtrar solo elementos disponibles.
          schema:
            type: boolean
      responses:
        '200':
          description: Catálogo obtenido correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemDetail'

    post:
      tags:
        - Inventario y Materiales
      summary: Ingreso/creación de un nuevo ítem al inventario
      description: (Solo Admin).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Ítem agregado con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '400':
          description: Datos incorrectos.
        '403':
          description: Prohibido. Permisos insuficientes.
...
  /api/reservas:
    get:
      tags:
        - Préstamos y Reservas
      summary: Lista los préstamos
      description: Admin ve todos, el alumno ve los propios. Permite filtrar por estado, fechas o disponibilidad.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Pendiente, Aprobado, Entregado, Devuelto, Rechazado]
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de préstamos obtenida.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanDetail'

    post:
      tags:
        - Préstamos y Reservas
      summary: Crea una nueva solicitud de reserva
      description: Endpoint de compatibilidad. Admin y bibliotecario pueden crear reservas para cualquier usuario; alumno y profesor solo pueden crear reservas propias.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanCreateLegacyRequest'
      responses:
        '201':
          description: Reserva creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanCreateLegacyResponse'
        '400':
          description: Validaciones fallidas.
        '404':
          description: Usuario o ítem no encontrado.
        '403':
          description: Prohibido. El usuario autenticado no puede crear reservas para otro usuario.
  /api/reservas/{reserva_id}:
    parameters:
      - name: reserva_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Préstamos y Reservas
      summary: Obtiene los detalles completos de un préstamo o solicitud
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detalles devueltos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDetail'
        '404':
          description: Préstamo no encontrado.
        '403':
          description: Prohibido. El usuario autenticado no puede consultar reservas ajenas.
...
  /api/penalizaciones:
    get:
      tags:
        - Penalizaciones
      summary: Lista general de penalizaciones activas e históricas
      description: Admin y bibliotecario pueden listar todas las penalizaciones; alumno y profesor solo pueden consultar las propias con usuario_id.
      security:
        - BearerAuth: []
      parameters:
        - name: usuario_id
          in: query
          required: false
          description: Identificador del usuario para filtrar penalizaciones.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Listado obtenido.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PenaltyDetail'
        '400':
          description: Parámetros de consulta inválidos.
        '403':
          description: Prohibido. El usuario autenticado no puede consultar penalizaciones ajenas.

    post:
      tags:
        - Penalizaciones
      summary: Genera una nueva penalización manual a un alumno
      description: Admin, bibliotecario o profesor pueden crear una penalización manual.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenaltyCreate'
      responses:
        '201':
          description: Penalización creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenaltyDetail'
        '400':
          description: Error en datos.
        '403':
          description: Prohibido.
...
  /api/normativas/:
    get:
      tags:
        - Normativas
      summary: Lista normativas con paginación HATEOAS
      parameters:
        - name: limit
          in: query
          required: false
          description: Cantidad máxima de registros a devolver. Default 10, máximo 100.
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Posición inicial del listado. Default 0.
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Normativas obtenidas con metadata y links HATEOAS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormativasResponse'
        '400':
          description: Parámetros de paginación inválidos.

  # ==========================================
  # 8. Reportes y Estadísticas (/api/reports)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        puntaje:
          type: integer
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
          nullable: true
...
    UserCreate:
      type: object
      required:
        - nombre
        - email
        - rol
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
        puntaje:
          type: integer
    UserUpdate:
      type: object
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
        puntaje:
          type: integer
...
    ItemDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre_art:
          type: string
        tipo:
          type: string
        seccion:
          type: string
        prestacion_maxima:
          type: integer
        stock:
          type: integer
        necesita_reparacion:
          type: boolean
...
    ItemCreate:
      type: object
      required:
        - nombre_art
        - tipo
        - seccion
        - prestacion_maxima
      properties:
        nombre_art:
          type: string
        tipo:
          type: string
        seccion:
          type: string
        prestacion_maxima:
          type: integer
        stock:
          type: integer
        necesita_reparacion:
          type: boolean
...
    LoanDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_usuario:
          type: integer
          format: int64
        id_reservado:
          type: integer
          format: int64
        estado_reserva:
          type: string
          enum: [pendiente, aprobado, entregado, devuelto, cancelado]
        fecha_retiro:
          type: string
          format: date-time
        fecha_regreso:
          type: string
          format: date-time
    LoanCompact:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_reservado:
          type: integer
          format: int64
        nombre_art:
          type: string
        estado_reserva:
          type: string
        fecha_retiro:
          type: string
          format: date-time
...
    LoanCreateLegacyRequest:
      type: object
      required:
        - usuario_id
        - articulo_id
      properties:
        usuario_id:
          type: integer
          format: int64
        articulo_id:
          type: integer
          format: int64
    LoanCreateLegacyResponse:
      type: object
      properties:
        message:
          type: string
        reserva_id:
          type: integer
          format: int64
        usuario_id:
          type: integer
          format: int64
        articulo_id:
          type: integer
          format: int64
...
    PenaltyDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        loanId:
          type: integer
          format: int64
          nullable: true
        reason:
          type: string
        status:
          type: string
          enum: [Activa, Levantada]
        severity:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
...